+++ deploy/095_set_Goerli_reward_pool.js
@@ -1,10 +1,12 @@
-const {Contract} = require("ethers");
+const {Contract, BigNumber} = require("ethers");
 const {ethers} = require("@nomiclabs/buidler");
 
 module.exports = async ({getNamedAccounts, deployments}) => {
-  const {read, execute, log} = deployments;
+  const {execute, log} = deployments;
 
   const {deployer, sandAdmin} = await getNamedAccounts();
+
+  const rewardAmount = BigNumber.from(2500000).mul("1000000000000000000");
 
   const rewardPoolAsDeployed = await deployments.get("GoerliSANDRewardPool");
   const rewardPoolABI = rewardPoolAsDeployed.abi;
@@ -18,7 +20,12 @@ module.exports = async ({getNamedAccounts, deployments}) => {
   log("setting sandAdmin as Reward Distribution");
   await rewardPoolAsDeployer.functions.setRewardDistribution(sandAdmin, {gasLimit: 1000000}).then((tx) => tx.wait());
 
+  // transfer SAND reward to rewardPoolAddress
+  log("transferring SAND reward");
+  await execute("Sand", {from: deployer, skipUnknownSigner: true}, "transfer", rewardPoolAddress, rewardAmount);
+
   log("notifying the Reward Amount");
-  await rewardPoolAsAdmin.functions.notifyRewardAmount(3000000, {gasLimit: 1000000}).then((tx) => tx.wait());
+  await rewardPoolAsAdmin.functions.notifyRewardAmount(rewardAmount, {gasLimit: 1000000}).then((tx) => tx.wait());
 };
+module.exports.skip = async () => true;
 module.exports.dependencies = ["GoerliSANDRewardPool", "Sand"];
